week2/day5/School_Management_System/back_end/requirements.txt
blinker==1.9.0
click==8.3.0
flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==3.0.10